ci(github): Rewrite the Docker build job
Replace the `docker-ort` workflow with the new `docker-build` workflow.
Instead of the old approach to build separate Docker images for the
different package managers, the new workflow uses the Docker registry
cache [1] to cache the layers in the GitHub container registry. This
dramatically simplifies the build process, as no additional logic for
building the Docker images is required anymore.

The workflow first builds the full Docker image and caches it to the
registry. The minimal image is built afterwards and can reuse all layers
but the last one from the previously populated cache.

[1]: https://docs.docker.com/build/cache/backends/registry/

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>