refactor(static-html): Extract constant for the rule violation table ID
Signed-off-by: Frank Viernau <frank_viernau@epam.com>

refactor(static-html): Align on the term "rule violation"

Consistently name it rule violations.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

refactor(static-html): Move the `p` tag out of `issueDescription()`

Simplify an upcoming change.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

test(reporters): Add analyzer issues to package references

Make the tests cover this case as well.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

refactor(model): Factor out `filterIssues()`

Prepare for re-use in an upcoming change.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

refactor(model): Expose sparate `getIssues()` for related ORT stages

Prepare for re-use in upcoming changes.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

feat(static-html): Re-design the summary of issues

Previously, there was a single issues table containing analyzer and
scanner issues. A single row was dedicated to a project or package and
contained all corresponding analyzer and scan issues. So, one of its
design goals must have been to make the set of issues corresponding
to any project or package easy to grasp / obvious. However, this
design also has a couple of shortcomings:

1. The view cannot be extended to show advisor issues, because that
   would require adding another column which would lead to too small
   columns on smaller or medium sized screens.
2. Issues may be shown multiple times if they relate to multiple
   projects. This can render the number of issues to be resolved, shown
   in the issue summary title, less meaningful.
3. Issues cannot be ordered by severity.
4. It's tricky to style a row WRT the severity, because each row may
   contain multiple issues differing in their severity.

This change changes the approach as follows:

1. Split-up the issues table into one table per stage, analyzer and
   scanner.
2. Make each row dedicated to a single issue.
3. Order issues by their severity.

This allows to easily go through the issues for each stage separately,
in order of the respective severity, which aligns with the natural
order in which issues are being worked on. The code becomes quite a bit
simpler, as the overall approach is simpler and the table becomes more
consistent with the rule violation summary table, which also has the
approach of one line per violation. While the link to the related
entries in the project tables is now missing, the new approach is
easily extensible by such links. For example, a column dedicated to
these links can be added. Finally, as the issue getters from
`OrtResult` are now used, issues with excluded affected paths are now
omitted from the summary view.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

test(reports): Add a couple of advisor issues to reporter test assets

Make the tests cover also unresolved errors, warnings and hints. While
at it, adjust the message of the resolved issue to be more speaking.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

feat(static-html): Add an advisor issue summary table

Show advisor issues analog to analyzer and scanner issues.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>